:recycle: 変換系メソッドを別ファイルに切り出してDIするように変更 #5
