+++ src/main/java/app/reminderappbackend/service/ReminderService.java
@@ -1,8 +1,5 @@
 package app.reminderappbackend.service;
 
-import java.time.LocalDateTime;
-import java.time.OffsetDateTime;
-import java.time.ZoneOffset;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -10,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import app.reminderappbackend.repository.ReminderRecord;
 import app.reminderappbackend.repository.ReminderRepository;
+import app.reminderappbackend.util.DataTypeConverter;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
@@ -20,6 +18,7 @@ import reminderapi.model.ReminderForm;
 public class ReminderService {
 
   private final ReminderRepository reminderRepository;
+  private final DataTypeConverter converter;
 
   /**
    * IDに紐づくリマインダーを取得するサービス
@@ -29,10 +28,9 @@ public class ReminderService {
    */
   public ReminderEntity findById(Long id) {
     var optRecord = reminderRepository.selectById(id);
-    var entity =
-        optRecord
-            .map(record -> toReminderEntity(record))
-            .orElseThrow(() -> new ReminderEntityNotFoundException(id));
+    var entity = optRecord
+      .map(record -> converter.toReminderEntity(record))
+      .orElseThrow(() -> new ReminderEntityNotFoundException(id));
 
     return entity;
   }
@@ -47,7 +45,7 @@ public class ReminderService {
   public List<ReminderEntity> findList(Integer limit, Long offset) {
     List<ReminderRecord> recordList = reminderRepository.selectList(limit, offset);
     var entityList = recordList.stream()
-      .map(record -> toReminderEntity(record))
+      .map(record -> converter.toReminderEntity(record))
       .collect(Collectors.toList());
 
     return entityList;
@@ -61,11 +59,11 @@ public class ReminderService {
    */
   public ReminderEntity create(@Valid ReminderForm form) {
     // form を record に詰めて、Repositoryへ渡す
-    var record = toReminderRecord(form);
+    var record = converter.toReminderRecord(form);
     reminderRepository.insert(record);
 
     // record to entity
-    var entity = toReminderEntity(record);
+    var entity = converter.toReminderEntity(record);
 
     return entity;
   }
@@ -100,38 +98,4 @@ public class ReminderService {
     reminderRepository.delete(id);
   }
 
-  private ReminderRecord toReminderRecord(ReminderForm form) {
-    return new ReminderRecord(
-        null,
-        form.getTitle(),
-        form.getDescription(),
-        form.getDueDate(),
-        form.getPriority(),
-        form.getIsCompleted(),
-        LocalDateTime.now(),
-        LocalDateTime.now());
-  }
-
-  private ReminderEntity toReminderEntity(ReminderRecord record) {
-    return new ReminderEntity(
-        record.getId(),
-        record.getTitle(),
-        record.getDescription(),
-        record.getDueDate(),
-        record.getPriority(),
-        record.isCompleted(),
-        toOffsetDateTime(record.getCreatedAt()),
-        toOffsetDateTime(record.getUpdatedAt()));
-  }
-
-  /**
-   * クライアントへデータを返却する際に、以下のカラムを<OffsetDateTime>型に変換する必要がある 理由は、OpenAPI
-   * Specificationで生成される<date-time>が<OffsetDateTime>型であるため - created_at - updated_at
-   *
-   * @param ldt
-   * @return OffsetDateTime
-   */
-  private OffsetDateTime toOffsetDateTime(LocalDateTime ldt) {
-    return ldt.atOffset(ZoneOffset.UTC);
-  }
 }
